requirements/base.txt
asgiref==3.2.7
attrs==19.3.0
behave==1.2.6
behave-django==1.3.0
certifi==2020.4.5.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
drf-yasg==1.17.1
glob2==0.7
idna==2.9
importlib-metadata==1.6.1
inflection==0.5.0
itypes==1.2.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
more-itertools==8.3.0
packaging==20.4
parse==1.15.0
parse-type==0.5.2
pkg-resources==0.0.0
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.7
pytz==2020.1
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
Unipath==1.1
uritemplate==3.0.1
urllib3==1.25.9
wcwidth==0.2.4
zipp==3.1.0
